fix #147, convert number to string in package common

+++ src/main/java/capstone/lip/landinformationportal/business/service/RealEstateAdjacentSegmentService.java
@@ -6,11 +6,10 @@
 package capstone.lip.landinformationportal.business.service;
 
 import capstone.lip.landinformationportal.business.repository.RealEstateAdjacentSegmentRepository;
+import capstone.lip.landinformationportal.business.repository.RealEstateRepository;
+import capstone.lip.landinformationportal.business.repository.SegmentOfStreetRepository;
 import capstone.lip.landinformationportal.business.service.Interface.IRealEstateAdjacentSegmentService;
-import capstone.lip.landinformationportal.business.validation.RealEstateValidation;
-import capstone.lip.landinformationportal.common.entity.RealEstate;
 import capstone.lip.landinformationportal.common.entity.RealEstateAdjacentSegment;
-import capstone.lip.landinformationportal.common.entity.compositekey.RealEstateAdjacentSegmentId;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,12 +24,24 @@ public class RealEstateAdjacentSegmentService implements IRealEstateAdjacentSegm
 
     @Autowired
     private RealEstateAdjacentSegmentRepository realEstateAdjacentSegmentRepository;
-
+    @Autowired
+    private RealEstateRepository realEstateRepoditory;
+    @Autowired
+    private SegmentOfStreetRepository segmentOfStreetRepository;
+    
     
 
     @Override
     public boolean delete(List<RealEstateAdjacentSegment> listRealEstateAdjacentSegment) {
     	try {
+    		if (listRealEstateAdjacentSegment == null) throw new Exception("null");
+    		if (listRealEstateAdjacentSegment.isEmpty()) throw new Exception("empty");
+    		for (RealEstateAdjacentSegment element: listRealEstateAdjacentSegment) {
+    			if (realEstateAdjacentSegmentRepository.findByIdSegmentOfStreetIdAndIdRealEstateId(
+    					element.getId().getSegmentOfStreetId(), element.getId().getRealEstateId())==null) {
+    				throw new Exception("Id not found");
+    			}
+    		}
     		realEstateAdjacentSegmentRepository.deleteAll(listRealEstateAdjacentSegment);
     		return true;
 		} catch (Exception e) {
@@ -43,6 +54,16 @@ public class RealEstateAdjacentSegmentService implements IRealEstateAdjacentSegm
     @Override
     public List<RealEstateAdjacentSegment> save(List<RealEstateAdjacentSegment> listReoAdjacentSegment) {
     	try {
+    		if (listReoAdjacentSegment == null) throw new Exception("null");
+    		if (listReoAdjacentSegment.isEmpty()) throw new Exception("empty");
+    		for (RealEstateAdjacentSegment element: listReoAdjacentSegment) {
+    			if (!realEstateRepoditory.existsById(element.getId().getRealEstateId())) {
+    				throw new Exception("Realestate id not found");
+    			}
+    			if (!segmentOfStreetRepository.existsById(element.getId().getSegmentOfStreetId())) {
+    				throw new Exception("Segment id not found");
+    			}
+    		}
     		return realEstateAdjacentSegmentRepository.saveAll(listReoAdjacentSegment);
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -54,6 +75,12 @@ public class RealEstateAdjacentSegmentService implements IRealEstateAdjacentSegm
     @Override
     public RealEstateAdjacentSegment save(RealEstateAdjacentSegment realEstateAdjacentSegment) {
     	try {
+    		if (!realEstateRepoditory.existsById(realEstateAdjacentSegment.getId().getRealEstateId())) {
+				throw new Exception("Realestate id not found");
+			}
+			if (!segmentOfStreetRepository.existsById(realEstateAdjacentSegment.getId().getSegmentOfStreetId())) {
+				throw new Exception("Segment id not found");
+			}
     		return realEstateAdjacentSegmentRepository.save(realEstateAdjacentSegment);
 		} catch (Exception e) {
 			e.printStackTrace();
